@@ -7,7 +7,7 @@ version: 1.0.0+1
 publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -24,9 +24,14 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.3
-  multi_image_picker:
-    git:
-      url: https://github.com/JosephNK/multi_image_picker.git
+
+  # Third Party
+  # multi_image_picker:
+  #   git:
+  #     url: https://github.com/JosephNK/multi_image_picker.git
+
+  # Google/Flutter/Dart
+  pedantic: ^1.11.0
 
 dev_dependencies:
   flutter_test:
